+++ packages/blade/src/components/Box/BaseBox/types/spacingTypes.ts
@@ -3,11 +3,20 @@ import type { DotNotationSpacingStringToken } from '~src/_helpers/types';
 
 type ArrayOfMaxLength4<T> = readonly [T?, T?, T?, T?];
 type SpaceUnits = 'px' | '%' | 'fr' | 'rem' | 'em' | 'vh' | 'vw';
+
 type SpacingValueType =
   | DotNotationSpacingStringToken
   | `${string}${SpaceUnits}`
   | `calc(${string})`
-  | 'auto';
+  | 'auto'
+  | `min(${string})`
+  | `max(${string})`
+  | 'none'
+  | 'initial'
+  | 'fit-content'
+  | 'max-content'
+  | 'min-content';
+
 /**
  * @IMPORTANT
  *