+++ packages/blade/src/components/Button/IconButton/types.ts
@@ -1,4 +1,5 @@
 import type { IconComponent } from '~components/Icons';
+import type { BladeCommonEvents } from '~components/types';
 import type { TestID } from '~src/_helpers/types';
 
 export type StyledIconButtonProps = {
@@ -7,4 +8,5 @@ export type StyledIconButtonProps = {
   size: 'medium' | 'large';
   contrast: 'low' | 'high';
   accessibilityLabel: string;
-} & TestID;
+} & TestID &
+  BladeCommonEvents;