+++ packages/blade/src/components/Icons/_Svg/Svg/Svg.native.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import styled from 'styled-components/native';
 import type { ReactElement } from 'react';
 import { Svg as SvgNative } from 'react-native-svg';
+import React from 'react';
 import type { SvgProps } from './types';
 import { makeAccessible } from '~utils';
 import { useStyledProps } from '~components/Box/styledProps';
@@ -10,14 +12,10 @@ const StyledSvg = styled(SvgNative)((props) => {
   return styledPropsCSSObject;
 });
 
-const Svg = ({
-  children,
-  height,
-  viewBox,
-  width,
-  fill,
-  ...styledProps
-}: SvgProps): ReactElement => {
+const _Svg: React.ForwardRefRenderFunction<any, SvgProps> = (
+  { children, height, viewBox, width, fill, ...styledProps },
+  ref,
+): ReactElement => {
   return (
     <StyledSvg
       {...makeAccessible({ hidden: true })}
@@ -25,11 +23,14 @@ const Svg = ({
       viewBox={viewBox}
       width={width}
       fill={fill}
+      ref={ref}
       {...styledProps}
     >
       {children}
     </StyledSvg>
   );
 };
+
+const Svg = React.forwardRef(_Svg);
 
 export default Svg;